+++ .changeset/nice-rockets-repeat.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-aptos": major
+---
+
+add initial Alpaca API integration for Aptos
