+++ libs/coin-modules/coin-aptos/.unimportedrc.json
@@ -1,16 +1,28 @@
 {
-  "ignoreUnresolved": ["follow-redirects", "form-data", "proxy-from-env"],
-  "ignoreUnused": ["@ledgerhq/devices", "invariant", "expect"],
+  "entry": [
+    "src/index.ts",
+    "src/api/index.ts",
+    "src/bridge/index.ts",
+    "src/network/index.ts",
+    "src/signer/index.ts"
+  ],
+  "ignoreUnused": [
+    "@ledgerhq/devices",
+    "@ledgerhq/types-cryptoassets",
+    "expect",
+    "invariant"
+  ],
   "ignoreUnimported": [
+    "src/bridge/bridge.fixture.ts",
     "src/bridge/deviceTransactionConfig.ts",
     "src/bridge/transaction.ts",
-    "src/bridge/bridge.fixture.ts",
+    "src/config.ts",
     "src/errors.ts",
-    "src/hw-signMessage.ts",
     "src/test/bot-specs.ts",
     "src/test/bridgeDatasetTest.ts",
     "src/test/cli.ts",
     "src/test/index.ts",
-    "src/test/speculos-deviceActions.ts"
+    "src/test/speculos-deviceActions.ts",
+    "src/types/assets.ts"
   ]
 }
