+++ libs/coin-modules/coin-aptos/jest.integ.config.js
@@ -1,0 +1,8 @@
+/** @type {import('ts-jest/dist/types').JestConfigWithTsJest} */
+module.exports = {
+  preset: "ts-jest",
+  testEnvironment: "node",
+  testRegex: ".integ.test.ts$",
+  testPathIgnorePatterns: ["lib/", "lib-es/"],
+  testTimeout: 60000,
+};
