+++ libs/coin-modules/coin-aptos/src/__tests__/bridge/getFeesForTransaction.test.ts
@@ -6,11 +6,11 @@ import {
   createFixtureTransactionWithSubAccount,
 } from "../../bridge/bridge.fixture";
 import * as getFeesForTransaction from "../../bridge/getFeesForTransaction";
-import { AptosAPI } from "../../api";
+import { AptosAPI } from "../../network";
 
 let simulateTransaction = jest.fn();
 
-jest.mock("../../api", () => {
+jest.mock("../../network", () => {
   return {
     AptosAPI: function () {
       return {
