+++ libs/coin-modules/coin-aptos/src/__tests__/bridge/prepareTransaction.test.ts
@@ -1,12 +1,12 @@
 import prepareTransaction from "../../bridge/prepareTransaction";
-import { AptosAPI } from "../../api";
+import { AptosAPI } from "../../network";
 import { getEstimatedGas } from "../../bridge/getFeesForTransaction";
 import { getMaxSendBalance } from "../../bridge/logic";
 import BigNumber from "bignumber.js";
 import type { Account } from "@ledgerhq/types-live";
 import type { Transaction } from "../../types";
 
-jest.mock("../../api");
+jest.mock("../../network");
 jest.mock("../../bridge/getFeesForTransaction");
 jest.mock("../../bridge/logic");
