+++ libs/coin-modules/coin-aptos/src/__tests__/bridge/synchronisation.test.ts
@@ -3,7 +3,7 @@ import { getCryptoCurrencyById } from "@ledgerhq/cryptoassets/index";
 import { Account, Operation, SyncConfig, TokenAccount } from "@ledgerhq/types-live";
 import { decodeTokenAccountId } from "@ledgerhq/coin-framework/account";
 import { emptyHistoryCache } from "@ledgerhq/coin-framework/account/index";
-import { AptosAPI } from "../../api";
+import { AptosAPI } from "../../network";
 import {
   getAccountShape,
   mergeSubAccounts,
@@ -35,11 +35,11 @@ jest.mock("@ledgerhq/coin-framework/account", () => {
 
 const mockedDecodeTokenAccountId = jest.mocked(decodeTokenAccountId);
 
+jest.mock("../../network");
+let mockedAptosAPI: jest.Mocked<any>;
+
 jest.mock("@ledgerhq/coin-framework/bridge/jsHelpers");
 jest.mock("invariant", () => jest.fn());
-
-jest.mock("../../api");
-let mockedAptosAPI: jest.Mocked<any>;
 
 jest.mock("../../bridge/logic");
