+++ libs/coin-modules/coin-aptos/src/bridge/broadcast.ts
@@ -1,16 +1,16 @@
 import type { Account, Operation, SignedOperation } from "@ledgerhq/types-live";
 import { patchOperationWithHash } from "@ledgerhq/coin-framework/operation";
-import { AptosAPI } from "../api";
+import { AptosAPI } from "../network";
 
-const broadcast = async ({
-  signedOperation,
-  account,
-}: {
+type broadcastFunc = {
   signedOperation: SignedOperation;
   account: Account;
-}): Promise<Operation> => {
+};
+
+const broadcast = async ({ signedOperation, account }: broadcastFunc): Promise<Operation> => {
   const { signature, operation } = signedOperation;
-  const hash = await new AptosAPI(account.currency.id).broadcast(signature);
+  const client = new AptosAPI(account.currency.id);
+  const hash = await client.broadcast(signature);
   return patchOperationWithHash(operation, hash);
 };
