+++ libs/coin-modules/coin-aptos/src/bridge/buildTransaction.ts
@@ -1,8 +1,8 @@
 import { InputEntryFunctionData, RawTransaction } from "@aptos-labs/ts-sdk";
 import type { Account, TokenAccount } from "@ledgerhq/types-live";
 import { findSubAccountById, isTokenAccount } from "@ledgerhq/coin-framework/account/index";
-import { AptosAPI } from "../api";
 import { APTOS_ASSET_ID, SUPPORTED_TOKEN_TYPES } from "../constants";
+import { AptosAPI } from "../network";
 import { normalizeTransactionOptions } from "./logic";
 import type { Transaction } from "../types";
