+++ libs/coin-modules/coin-aptos/src/network/index.ts
@@ -9,10 +9,11 @@ import {
   RawTransaction,
   SimpleTransaction,
 } from "@aptos-labs/ts-sdk";
-import { SignerContext } from "@ledgerhq/coin-framework/lib/signer";
+import { SignerContext } from "@ledgerhq/coin-framework/signer";
 import { Account } from "@ledgerhq/types-live";
 import { AptosSigner } from "../types";
 import { sha3_256 as sha3Hash } from "@noble/hashes/sha3";
+export * from "./client";
 
 export async function signTransaction(
   signerContext: SignerContext<AptosSigner>,
