+++ libs/coin-modules/coin-aptos/src/types/assets.ts
@@ -1,0 +1,3 @@
+import type { Asset } from "@ledgerhq/coin-framework/api/types";
+
+export type AptosAsset = Asset;
